+++ src/components/Command/Command.tsx
@@ -1,0 +1,11 @@
+export {
+  Command,
+  CommandDialog,
+  CommandInput,
+  CommandList,
+  CommandEmpty,
+  CommandGroup,
+  CommandItem,
+  CommandSeparator,
+  CommandShortcut,
+} from "../../primitives/command"