+++ src/components/Form/Form.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useCallback, useMemo, type ReactNode } from "react";
 import {
   useForm,